Adding everything to the repository?
I don't know how github works??

+++ Projectiles/HarmonyOrb.cs
@@ -7,6 +7,7 @@ namespace Necromancy.Projectiles
 {
 	public class HarmonyOrb : ModProjectile
 	{
+        // stays over a targeted player and heals them
         public override void SetStaticDefaults()
         {
             DisplayName.SetDefault("Orb of Harmony");
@@ -25,39 +26,30 @@ namespace Necromancy.Projectiles
 
 		public override void AI()
 		{
-            Player owner = Main.player[projectile.owner];
-            Player target = Main.player[(int)projectile.ai[0]];
-            Lighting.AddLight(projectile.Center, new Vector3(1f, 1f, 0.4f));
-            if (!target.dead && owner.channel && Vector2.Distance(projectile.Center, owner.Center) < 1000f)
+            if (Main.myPlayer == projectile.owner)
             {
-                if (projectile.timeLeft <= 1)
+                projectile.netUpdate = true;
+                Player owner = Main.player[projectile.owner];
+                Player target = Main.player[(int)projectile.ai[0]];
+                Lighting.AddLight(projectile.Center, new Vector3(1f, 1f, 0.4f));
+                if (!target.dead && owner.channel && Vector2.Distance(projectile.Center, owner.Center) < 1000f)
                 {
-                    owner.statLife--;
-                    if (owner.statLife <= 0)
+                    if (projectile.timeLeft <= 1)
                     {
-                        Terraria.DataStructures.PlayerDeathReason damageSource;
-                        switch (Main.rand.Next(3))
-                        {
-                            case 0:
-                                damageSource = Terraria.DataStructures.PlayerDeathReason.ByCustomReason(owner.name + " ran out of blood.");
-                                break;
-                            case 1:
-                                damageSource = Terraria.DataStructures.PlayerDeathReason.ByCustomReason(owner.name + " couldn't handle the power.");
-                                break;
-                            default:
-                                damageSource = Terraria.DataStructures.PlayerDeathReason.ByCustomReason(owner.name + " didn't watch their health bar.");
-                                break;
-                        }
-                        owner.KillMe(damageSource, 5, -owner.direction);
+                        Necromancy.DrainLife(owner, 1);
+                        projectile.timeLeft = 4;
                     }
-                    projectile.timeLeft = 4;
+                    projectile.Center = new Vector2(target.Center.X, target.position.Y - 24f);
+                    target.AddBuff(mod.BuffType<Buffs.HarmonyOrb>(), 2);
+
+                    Vector2 dustPos = projectile.Center + Main.rand.NextVector2Circular(projectile.width / 2, projectile.height / 2);
+                    Dust d = Dust.QuickDust(dustPos, new Color(1f, 1f, 0.5f));
+                    d.velocity = 0.08f * (target.Center - d.position) + target.velocity;
                 }
-                projectile.Center = new Vector2(target.Center.X, target.position.Y - 24f);
-                target.AddBuff(mod.BuffType<Buffs.HarmonyOrb>(), 2);
-            }
-            else
-            {
-                projectile.Kill();
+                else
+                {
+                    projectile.Kill();
+                }
             }
         }
     }
